Removed etransforms code. BlockTriDiagonalTransform is now only used in tests.
